✨ feat(workspace): add Dashboard with WorkspaceCard and CreateModal
